+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed a bug where some string functions would incorrectly handle newlines when
   iterating over graphemes in older JavaScript environments that do not have the
   `Intl.Segmenter` class.
+- Added function `iterator.tap` for adding side effects to iterator pipelines.
 
 ## v0.25.0 - 2022-11-19